refactor(client): implement Pinia and restructure application architecture
