+++ client/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@app/shared": "file:../shared",
+    "pinia": "^3.0.4",
     "socket.io-client": "^4.8.3",
     "vue": "^3.5.26"
   },
